update jupyter server entrypoint
The entrypoint for jupyter server appears to have changed to be prefixed
with an underscore, so this change updates the entrypoint, but leaves
the existing name for legacy users. The documentation is also updated to
show how to configure this extension if using `jupyter server` (JupyterLab).

See: https://jupyter-server.readthedocs.io/en/latest/developers/extensions.html#migrating-an-extension-to-use-jupyter-server